📦 NEW: Add portfolio
