+++ packages/web/constants/Stack.js
@@ -21,8 +21,6 @@ export const TECH = {
     'Redux',
     'Heroku',
     'MongoDB',
-    'Markdown',
-    'Styled-Components',
   ],
 }
 
@@ -49,6 +47,33 @@ export const SOCIAL = {
     {
       platform: 'Facebook',
       link: `https://www.facebook.com/${META.social.facebook}`,
+    },
+  ],
+}
+
+export const PROJECTS = {
+  title: 'Projects I have worked on',
+  data: [
+    {
+      image: 'https://i.ibb.co/dPRzYZv/pkgreview-2.jpg',
+      name: 'pkgreview.dev',
+      link: 'https://pkgreview.dev',
+      description:
+        'A lot of people find it very difficult to decide what NPM Package to use. pkgreview.dev helps developers in the JavaScript Community review and analyze each module before they start using it.',
+    },
+    {
+      image: 'https://i.ibb.co/fd2rF97/favicon.jpg',
+      name: 'JoinMyCall',
+      link: 'https://joinmycall.now.sh',
+      description:
+        'This app makes your colleagues take a connection/camera/mic test before joining a call. Hosts can create infinite JoinMyCall links to invite guests in meeting. Guests can either skip or do the connection test.',
+    },
+    {
+      image: 'https://kumar.now.sh/static/prebuilt/images/bulkmail.gif',
+      name: 'bulk-mail-cli',
+      link: 'https://bulkmail.now.sh',
+      description:
+        'A command line interface (cli) application that sends automated and dynamic emails to the list.',
     },
   ],
 }
