+++ packages/web/styles/base.css
@@ -434,6 +434,7 @@ form input:focus {
   margin-bottom: 20px;
   box-shadow: inset 0 -16px 0 var(--highlight-color);
   display: inline-block;
+  /* color: var(--heading-color); */
 }
 
 .uses-list li:not(.head) {
